Reworked weight system
- Weights now sum to 1.0.
- Changed weight of stress updates to 0.0
- This has the effect of loading up all stress before the world
  actually loads.

+++ src/main/kotlin/graveler/action/Action.kt
@@ -1,7 +1,7 @@
 package graveler.action
 
 interface Action {
-  val weight: Int get() = 1
+  val weight: Double get() = 1.0
 
   fun apply(context: ActionContext)
 }
